feat(apis/pricing): incomplete gql implementation

+++ apis/pricing/src/models/external_data_transfer_request.rs
@@ -1,10 +1,11 @@
+use juniper::GraphQLInputObject;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+#[derive(Debug, Serialize, Deserialize, Clone, GraphQLInputObject)]
 #[serde(rename_all = "kebab-case")]
 pub struct ExternalDataTransferRequest {
     pub from_region_code: Option<String>,
-    pub start_range: Option<i64>,
+    pub start_range: Option<i32>,
     pub sort_by: Option<String>,
     pub sort_order: Option<String>,
 }
